@@ -27,13 +27,13 @@
         "psr-4": {"fivefilters\\Readability\\Test\\": "test"}
     },
     "require": {
-        "php": ">=7.3.0",
+        "php": ">=7.4.0",
         "ext-dom": "*",
         "ext-xml": "*",
         "ext-mbstring": "*",
         "psr/log": "^1.0",
         "masterminds/html5": "^2.0",
-        "league/uri": "^6.4"
+        "league/uri": "~6.7.2"
     },
     "require-dev": {
         "phpunit/phpunit": "^9",
