@@ -3,7 +3,7 @@
         'name' => 'fivefilters.org/full-text-rss',
         'pretty_version' => 'dev-main',
         'version' => 'dev-main',
-        'reference' => '90b445223e27c0ada1f88983f1b4a0ef5063801d',
+        'reference' => '21710f18e63c3b256e5cc6cfe853bcb8feab93f3',
         'type' => 'project',
         'install_path' => __DIR__ . '/../',
         'aliases' => array(),
@@ -13,27 +13,25 @@
         'fivefilters.org/full-text-rss' => array(
             'pretty_version' => 'dev-main',
             'version' => 'dev-main',
-            'reference' => '90b445223e27c0ada1f88983f1b4a0ef5063801d',
+            'reference' => '21710f18e63c3b256e5cc6cfe853bcb8feab93f3',
             'type' => 'project',
             'install_path' => __DIR__ . '/../',
             'aliases' => array(),
             'dev_requirement' => false,
         ),
         'fivefilters/readability.php' => array(
-            'pretty_version' => 'dev-master',
-            'version' => 'dev-master',
-            'reference' => 'ec7fb4d8504118d918699d5df8cd07c91af2f9b3',
+            'pretty_version' => 'v3.1.2',
+            'version' => '3.1.2.0',
+            'reference' => '973fbb0f9c2335153c796dc5dce60f60e3adc347',
             'type' => 'library',
             'install_path' => __DIR__ . '/../fivefilters/readability.php',
-            'aliases' => array(
-                0 => '9999999-dev',
-            ),
+            'aliases' => array(),
             'dev_requirement' => false,
         ),
         'league/uri' => array(
-            'pretty_version' => '6.5.0',
-            'version' => '6.5.0.0',
-            'reference' => 'c68ca445abb04817d740ddd6d0b3551826ef0c5a',
+            'pretty_version' => '6.7.2',
+            'version' => '6.7.2.0',
+            'reference' => 'd3b50812dd51f3fbf176344cc2981db03d10fe06',
             'type' => 'library',
             'install_path' => __DIR__ . '/../league/uri',
             'aliases' => array(),
@@ -76,9 +74,9 @@
             'dev_requirement' => false,
         ),
         'simplepie/simplepie' => array(
-            'pretty_version' => '1.6.0',
-            'version' => '1.6.0.0',
-            'reference' => '2bdbc51ed1010941c9c5f2cddca433e79665bfe1',
+            'pretty_version' => '1.8.0',
+            'version' => '1.8.0.0',
+            'reference' => '65b095d87bc00898d8fa7737bdbcda93a3fbcc55',
             'type' => 'library',
             'install_path' => __DIR__ . '/../simplepie/simplepie',
             'aliases' => array(),
