@@ -1,14 +1,13 @@
 <?php
+
+declare(strict_types=1);
 /**
  * SimplePie
  *
  * A PHP-Based RSS and Atom Feed Framework.
  * Takes the hard work out of managing a complete RSS/Atom solution.
  *
- * Please note: This file is automatically generated by a build script. The
- * full original source is always available from http://simplepie.org/
- *
- * Copyright (c) 2004-2016, Ryan Parman, Sam Sneddon, Ryan McCue, and contributors
+ * Copyright (c) 2004-2022, Ryan Parman, Sam Sneddon, Ryan McCue, and contributors
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without modification, are
@@ -46,10 +45,76 @@
 
 namespace SimplePie\Cache;
 
-interface_exists('SimplePie_Cache_Base');
+/**
+ * Base for cache objects
+ *
+ * Classes to be used with {@see \SimplePie\Cache::register()} are expected
+ * to implement this interface.
+ *
+ * @package SimplePie
+ * @subpackage Caching
+ * @deprecated since SimplePie 1.8.0, use "Psr\SimpleCache\CacheInterface" instead
+ */
+interface Base
+{
+    /**
+     * Feed cache type
+     *
+     * @var string
+     */
+    public const TYPE_FEED = 'spc';
 
-if (\false) {
-	interface Base extends \SimplePie_Cache_Base
-	{
-	}
+    /**
+     * Image cache type
+     *
+     * @var string
+     */
+    public const TYPE_IMAGE = 'spi';
+
+    /**
+     * Create a new cache object
+     *
+     * @param string $location Location string (from SimplePie::$cache_location)
+     * @param string $name Unique ID for the cache
+     * @param Base::TYPE_FEED|Base::TYPE_IMAGE $type Either TYPE_FEED for SimplePie data, or TYPE_IMAGE for image data
+     */
+    public function __construct($location, $name, $type);
+
+    /**
+     * Save data to the cache
+     *
+     * @param array|\SimplePie\SimplePie $data Data to store in the cache. If passed a SimplePie object, only cache the $data property
+     * @return bool Successfulness
+     */
+    public function save($data);
+
+    /**
+     * Retrieve the data saved to the cache
+     *
+     * @return array Data for SimplePie::$data
+     */
+    public function load();
+
+    /**
+     * Retrieve the last modified time for the cache
+     *
+     * @return int Timestamp
+     */
+    public function mtime();
+
+    /**
+     * Set the last modified time to the current time
+     *
+     * @return bool Success status
+     */
+    public function touch();
+
+    /**
+     * Remove the cache
+     *
+     * @return bool Success status
+     */
+    public function unlink();
 }
+
+class_alias('SimplePie\Cache\Base', 'SimplePie_Cache_Base');
