@@ -1,14 +1,13 @@
 <?php
+
+declare(strict_types=1);
 /**
  * SimplePie
  *
  * A PHP-Based RSS and Atom Feed Framework.
  * Takes the hard work out of managing a complete RSS/Atom solution.
  *
- * Please note: This file is automatically generated by a build script. The
- * full original source is always available from http://simplepie.org/
- *
- * Copyright (c) 2004-2016, Ryan Parman, Sam Sneddon, Ryan McCue, and contributors
+ * Copyright (c) 2004-2022, Ryan Parman, Sam Sneddon, Ryan McCue, and contributors
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without modification, are
@@ -46,10 +45,141 @@
 
 namespace SimplePie\Cache;
 
-class_exists('SimplePie_Cache_Memcache');
+use Memcache as NativeMemcache;
 
-if (\false) {
-	class Memcache extends \SimplePie_Cache_Memcache
-	{
-	}
+/**
+ * Caches data to memcache
+ *
+ * Registered for URLs with the "memcache" protocol
+ *
+ * For example, `memcache://localhost:11211/?timeout=3600&prefix=sp_` will
+ * connect to memcache on `localhost` on port 11211. All tables will be
+ * prefixed with `sp_` and data will expire after 3600 seconds
+ *
+ * @package SimplePie
+ * @subpackage Caching
+ * @uses Memcache
+ * @deprecated since SimplePie 1.8.0, use implementation of "Psr\SimpleCache\CacheInterface" instead
+ */
+class Memcache implements Base
+{
+    /**
+     * Memcache instance
+     *
+     * @var Memcache
+     */
+    protected $cache;
+
+    /**
+     * Options
+     *
+     * @var array
+     */
+    protected $options;
+
+    /**
+     * Cache name
+     *
+     * @var string
+     */
+    protected $name;
+
+    /**
+     * Create a new cache object
+     *
+     * @param string $location Location string (from SimplePie::$cache_location)
+     * @param string $name Unique ID for the cache
+     * @param Base::TYPE_FEED|Base::TYPE_IMAGE $type Either TYPE_FEED for SimplePie data, or TYPE_IMAGE for image data
+     */
+    public function __construct($location, $name, $type)
+    {
+        $this->options = [
+            'host' => '127.0.0.1',
+            'port' => 11211,
+            'extras' => [
+                'timeout' => 3600, // one hour
+                'prefix' => 'simplepie_',
+            ],
+        ];
+        $this->options = array_replace_recursive($this->options, \SimplePie\Cache::parse_URL($location));
+
+        $this->name = $this->options['extras']['prefix'] . md5("$name:$type");
+
+        $this->cache = new NativeMemcache();
+        $this->cache->addServer($this->options['host'], (int) $this->options['port']);
+    }
+
+    /**
+     * Save data to the cache
+     *
+     * @param array|\SimplePie\SimplePie $data Data to store in the cache. If passed a SimplePie object, only cache the $data property
+     * @return bool Successfulness
+     */
+    public function save($data)
+    {
+        if ($data instanceof \SimplePie\SimplePie) {
+            $data = $data->data;
+        }
+        return $this->cache->set($this->name, serialize($data), MEMCACHE_COMPRESSED, (int) $this->options['extras']['timeout']);
+    }
+
+    /**
+     * Retrieve the data saved to the cache
+     *
+     * @return array Data for SimplePie::$data
+     */
+    public function load()
+    {
+        $data = $this->cache->get($this->name);
+
+        if ($data !== false) {
+            return unserialize($data);
+        }
+        return false;
+    }
+
+    /**
+     * Retrieve the last modified time for the cache
+     *
+     * @return int Timestamp
+     */
+    public function mtime()
+    {
+        $data = $this->cache->get($this->name);
+
+        if ($data !== false) {
+            // essentially ignore the mtime because Memcache expires on its own
+            return time();
+        }
+
+        return false;
+    }
+
+    /**
+     * Set the last modified time to the current time
+     *
+     * @return bool Success status
+     */
+    public function touch()
+    {
+        $data = $this->cache->get($this->name);
+
+        if ($data !== false) {
+            return $this->cache->set($this->name, $data, MEMCACHE_COMPRESSED, (int) $this->options['extras']['timeout']);
+        }
+
+        return false;
+    }
+
+    /**
+     * Remove the cache
+     *
+     * @return bool Success status
+     */
+    public function unlink()
+    {
+        return $this->cache->delete($this->name, 0);
+    }
 }
+
+class_alias('SimplePie\Cache\Memcache', 'SimplePie_Cache_Memcache');
