@@ -1,14 +1,13 @@
 <?php
+
+declare(strict_types=1);
 /**
  * SimplePie
  *
  * A PHP-Based RSS and Atom Feed Framework.
  * Takes the hard work out of managing a complete RSS/Atom solution.
  *
- * Please note: This file is automatically generated by a build script. The
- * full original source is always available from http://simplepie.org/
- *
- * Copyright (c) 2004-2016, Ryan Parman, Sam Sneddon, Ryan McCue, and contributors
+ * Copyright (c) 2004-2022, Ryan Parman, Sam Sneddon, Ryan McCue, and contributors
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without modification, are
@@ -46,10 +45,84 @@
 
 namespace SimplePie;
 
-class_exists('SimplePie_Copyright');
+/**
+ * Manages `<media:copyright>` copyright tags as defined in Media RSS
+ *
+ * Used by {@see \SimplePie\Enclosure::get_copyright()}
+ *
+ * This class can be overloaded with {@see \SimplePie\SimplePie::set_copyright_class()}
+ *
+ * @package SimplePie
+ * @subpackage API
+ */
+class Copyright
+{
+    /**
+     * Copyright URL
+     *
+     * @var string
+     * @see get_url()
+     */
+    public $url;
 
-if (\false) {
-	class Copyright extends \SimplePie_Copyright
-	{
-	}
+    /**
+     * Attribution
+     *
+     * @var string
+     * @see get_attribution()
+     */
+    public $label;
+
+    /**
+     * Constructor, used to input the data
+     *
+     * For documentation on all the parameters, see the corresponding
+     * properties and their accessors
+     */
+    public function __construct($url = null, $label = null)
+    {
+        $this->url = $url;
+        $this->label = $label;
+    }
+
+    /**
+     * String-ified version
+     *
+     * @return string
+     */
+    public function __toString()
+    {
+        // There is no $this->data here
+        return md5(serialize($this));
+    }
+
+    /**
+     * Get the copyright URL
+     *
+     * @return string|null URL to copyright information
+     */
+    public function get_url()
+    {
+        if ($this->url !== null) {
+            return $this->url;
+        }
+
+        return null;
+    }
+
+    /**
+     * Get the attribution text
+     *
+     * @return string|null
+     */
+    public function get_attribution()
+    {
+        if ($this->label !== null) {
+            return $this->label;
+        }
+
+        return null;
+    }
 }
+
+class_alias('SimplePie\Copyright', 'SimplePie_Copyright');
