@@ -1,14 +1,13 @@
 <?php
+
+declare(strict_types=1);
 /**
  * SimplePie
  *
  * A PHP-Based RSS and Atom Feed Framework.
  * Takes the hard work out of managing a complete RSS/Atom solution.
  *
- * Please note: This file is automatically generated by a build script. The
- * full original source is always available from http://simplepie.org/
- *
- * Copyright (c) 2004-2016, Ryan Parman, Sam Sneddon, Ryan McCue, and contributors
+ * Copyright (c) 2004-2022, Ryan Parman, Sam Sneddon, Ryan McCue, and contributors
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without modification, are
@@ -46,10 +45,15 @@
 
 namespace SimplePie;
 
-class_exists('SimplePie_Exception');
+use Exception as NativeException;
 
-if (\false) {
-	class Exception extends \SimplePie_Exception
-	{
-	}
+/**
+ * General SimplePie exception class
+ *
+ * @package SimplePie
+ */
+class Exception extends NativeException
+{
 }
+
+class_alias('SimplePie\Exception', 'SimplePie_Exception');
