@@ -1,14 +1,13 @@
 <?php
+
+declare(strict_types=1);
 /**
  * SimplePie
  *
  * A PHP-Based RSS and Atom Feed Framework.
  * Takes the hard work out of managing a complete RSS/Atom solution.
  *
- * Please note: This file is automatically generated by a build script. The
- * full original source is always available from http://simplepie.org/
- *
- * Copyright (c) 2004-2016, Ryan Parman, Sam Sneddon, Ryan McCue, and contributors
+ * Copyright (c) 2004-2022, Ryan Parman, Sam Sneddon, Ryan McCue, and contributors
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without modification, are
@@ -46,10 +45,84 @@
 
 namespace SimplePie;
 
-class_exists('SimplePie_Rating');
+/**
+ * Handles `<media:rating>` or `<itunes:explicit>` tags as defined in Media RSS and iTunes RSS respectively
+ *
+ * Used by {@see \SimplePie\Enclosure::get_rating()} and {@see \SimplePie\Enclosure::get_ratings()}
+ *
+ * This class can be overloaded with {@see \SimplePie\SimplePie::set_rating_class()}
+ *
+ * @package SimplePie
+ * @subpackage API
+ */
+class Rating
+{
+    /**
+     * Rating scheme
+     *
+     * @var string
+     * @see get_scheme()
+     */
+    public $scheme;
 
-if (\false) {
-	class Rating extends \SimplePie_Rating
-	{
-	}
+    /**
+     * Rating value
+     *
+     * @var string
+     * @see get_value()
+     */
+    public $value;
+
+    /**
+     * Constructor, used to input the data
+     *
+     * For documentation on all the parameters, see the corresponding
+     * properties and their accessors
+     */
+    public function __construct($scheme = null, $value = null)
+    {
+        $this->scheme = $scheme;
+        $this->value = $value;
+    }
+
+    /**
+     * String-ified version
+     *
+     * @return string
+     */
+    public function __toString()
+    {
+        // There is no $this->data here
+        return md5(serialize($this));
+    }
+
+    /**
+     * Get the organizational scheme for the rating
+     *
+     * @return string|null
+     */
+    public function get_scheme()
+    {
+        if ($this->scheme !== null) {
+            return $this->scheme;
+        }
+
+        return null;
+    }
+
+    /**
+     * Get the value of the rating
+     *
+     * @return string|null
+     */
+    public function get_value()
+    {
+        if ($this->value !== null) {
+            return $this->value;
+        }
+
+        return null;
+    }
 }
+
+class_alias('SimplePie\Rating', 'SimplePie_Rating');
